OHRM5X-1251 - Renamed files and namespaces
