+++ installer/config/routes.yaml
@@ -74,10 +74,10 @@ api_system_check:
 
 api_upgrader_start_registration:
   path: /upgrader/api/send-data/upgrader-start
-  controller: OrangeHRM\Installer\Controller\Upgrader\Api\RegisterDataOnStartAPI::handle
+  controller: OrangeHRM\Installer\Controller\Upgrader\Api\InitialDataRegistrationAPI::handle
   methods: [ POST ]
 
 api_upgrader_sucess_registration:
   path: /upgrader/api/send-data/upgrader-success
-  controller: OrangeHRM\Installer\Controller\Upgrader\Api\RegisterDataOnSuccessAPI::handle
+  controller: OrangeHRM\Installer\Controller\Upgrader\Api\SuccessDataRegistrationAPI::handle
   methods: [ POST ]
